pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.13.0",
    "chromadb>=1.5.5",
    "faiss-cpu>=1.13.2",
    "ipykernel>=7.2.0",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.23",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "numpy>=2.4.3",
    "openai>=2.30.0",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.9.2",
    "requests>=2.33.0",
    "sentence-transformers>=5.3.0",
    "torch>=2.11.0",
    "transformers>=5.4.0",
]
